Fix stack overflow in jsond_encode()
The JSON encoder is recursive, and it is far from easy to make it
iterative. Add a cheap stack limit check to prevent a segfault.
This uses the PHP_JSOND_ERROR_DEPTH error code that already talks about
the stack depth. Previously this was only used for the $depth argument.